+++ hotel_management_odoo/__manifest__.py
@@ -21,7 +21,7 @@
 #############################################################################
 {
     'name': 'Hotel Management',
-    'version': '19.0.1.0.3',
+    'version': '19.0.1.0.4',
     'category': 'Industries',
     'summary': """A complete Hotel Management System that cover all areas of 
      Hotel services""" ,